+++ pylj/md.py
@@ -28,7 +28,7 @@ def initialise(
         Initial temperature of the particles, in Kelvin.
     box_length: float
         Length of a single dimension of the simulation square, in Angstrom.
-    init_conf: string, optional
+    init_conf: string
         The way that the particles are initially positioned. Should be one of:
         - 'square'
         - 'random'